Adds results of re-execution on an i7 machine
Exact model of CPU and other details are in file facts/nibbler

MKL version is unavailable in the docker image, so those files are
removed. Also, some results (trsm_blis_nofus_fpu_mxn.m and
trsm_blis_nofus_sse_4x4.m) for BLIS are not being generated (need to
investigate why) so we also remove them to avoid confusions.

For those results that are valid (i.e. re-execution generates new
results), we can see that in most cases BLIS performs poorly. We need to
investigate if it is because the docker image was compiled using BLIS's
`reference` compiler configuration.