Add support for providing string timeout values
Support for human friendly timeout values in string has been added.
A function has been added in utils.py which parses the timeout
value which can be given as 20s or 10m etc. --timeout option now
takes integer or string as string and it is converted to seconds
accordingly.

Fixes #228.