add a no-publish flag and change the workflow
The workflow is changed now such that if the no-publish flag is
passed, a new archive is created and uploaded. If the flag is not
published, at first it is checked if a non-published upload
already exists. If yes, it will be passed otherwise a new archive
is created, uploaded and published.

fixes #312