Fixes #332. Add popper matrix subcommand (#364)
* Add popper matrix subcommand

popper matrix subcommand can be used to define an execution
matrix for a pipeline using environment variables. It is used as
`popper matrix pipeline --add key1=val1` to add a key and `popper
matrix pipeline --rm key1` to remove a key. Fixes #332.

* Modify utils.py and cmd_init.py

1. Added vars as an empty list during pipeline initialisation.
2. Added a utility function for reading pipeline configuration.
3. Modified the update_config function so that stages, envs, path
and vars are named arguments and have default values set for them.

* Rename subcommand matrix to env-vars

1. matrix['vars'] is chnaged to vars in the .popper.yml.
2. vars is now changed from dict to a list of dicts.
3. Addition and removal now handles the change as requested, i.e.,
add an element in the list of vars or remove an element from the
list.

* Make requested changes

* Add tests for popper env-vars

* Add test-env-vars to .popper.yml

* Change permissions for test-env-vars.sh file

* Fix update_config method in utils.py

fixes #332