adds transparent execution of workflow in singularity (#699)
fixes #697 - adds a --runtime flag that allows to provide 'singularity' or 'docker' (default). When singularity is given, the workflow is executed in singularity containers.

fixes #670 - correctly configures the bindings for HOME and Github Workflow environment variables

fixes #678 - changes the implementation of '--parallel' so that it uses a process pool instead of a thread one.