adds support for having variables in a workflow file (#758)
Adds ability to have variables in a workflow file. The format is `$_<VAR>`. The value of a variable is defined by a new --substitution flag of 'popper run' command. An `--allow-loose` flag is also added that determines whether to _not_ error out if one or more variables are not defined.

fixes #753 

Co-authored-by: Ivo Jimenez <ivo.jimenez@gmail.com>