major refactoring of codebase (#792)
Refactors WorkflowRunner and StepRunner. The central change is the move from having one instance of a step runner for every step in a workflow to having singleton runners (one for each engine). Unit testing is greatly simplified as well, making it easier to test code and increase coverage. Some functionality has been left out (mainly parallel execution), with the goal of bring it back later in refactored form. A bunch of small miscellaneous changes are also included.