adds PopperTest class and `popper versoin` test (#803)
Adds a PopperTest class that other unit tests can derive from in order
to make use of generic testing functionality. Currently this contains a
`mk_repo()` function that takes care of initializing a git repo with a
`README` file in it.

In addition, a test for the `popper version` command has been added
that makes use of Click's unit testing features that allow to test CLI
functionality. This will eventually replace the pipeline in the `ci/` folder.