Obtain branch name in CI environment
Adds support to obtain the name of the git branch associated to the
current commit in CI environments. In this scenarios, the commit is
checked out in detached HEAD state, which does not allow to obtain the
name of the branch via Git. CI services such as Gitlab, Jenkins, Travis,
etc. define a variable name that contains the name of the branch, which
is what this code looks for in order to define the CI service-agnostic
GIT_BRANCH variable.