Reimplement workflow parser (drops HCL syntax)
Reimplement parser assuming no backwards-compatibility for HCL. This new
implementation makes use of pykwalify and python-box to significantly
simplify the parsing and validation of workflow specification, as all of
this is done by these libraries instead of the Popper codebase.

One breaking change is that now step attributes 'runs', 'args', and
'secrets' are now assumed to be list of strings, whereas before these
could be either a string or list of strings.