Add new installer
From time to time, new users experience issues that stem from installing
the Python package in unforeseen environments. This results in poor user
experience. In order to solve this, this commit introduces a sh-based
installer that creates a script wrapper for running popper via Docker.

Since Docker is a requirement already, this has a higher likelihood of
working, once a Docker installation has been successful.