cli: fixes generation of image tag for docker runner
When a step references a local dockerfile, the tag of the image was
being assigned to the GIT_TAG of the underlying git project. When a
workflow gets executed in a folder that is not part of a repo, an empty
tag (with a `{}` string) was being generated and this caused an error:

  [1] docker build popper_1_step:{} /private/tmp/example/./foo
  ERROR: 500 Server Error: Internal Server Error ("invalid reference format")

With this commit adds logic to generate "na" as the tag for the image,
when running inside a "gitless" folder.

fixes #886