cli: add support for podman (#903)
This commit adds support for Podman. Since there is no stable python API yet, this implementation is done by opening subprocesses and executing the podman command. In addition, an issue with the latest podman versions (2.0.x series) is affecting one of the tests (see https://github.com/containers/podman/issues/7150), so this is being disabled for now.

fixes #704