cli: modify way in which empty git repos are identified (#916)
Small tweak on how empty Git repos are identified. Uses `git rev-list 
-n 1 -all` to check if there are commits in the currently checked out 
branch.

The value of the `GIT_TAG` environment variable when the currently 
checked out commit is not tagged is now an empty string, instead of 
`None`.

A small fix to the try-catch statement in `HostRunner._exec_cmd` 
function is also included.